I broke up routes.py into multiple files for simplicity and removed deprecated features(test functions). I also added debug information to monopoly for chromebooks, since Inspect is disabled.
