+++ templates/monopoly.html
@@ -5,6 +5,7 @@
 {% block content %}
 
 <link rel="stylesheet" href="{{url_for('static', filename='css/monopoly.css')}}" />
+<h3 id="console"></h3>
 <main>
     <div class="table stop">
         <div class="frame">
